Split Hashicorp across modules
