+++ ethsigner/app/src/main/java/tech/pegasys/ethsigner/FileBasedTransactionSignerCommand.java
@@ -12,6 +12,7 @@
  */
 package tech.pegasys.ethsigner;
 
+import tech.pegasys.ethsigner.core.signing.TransactionSigner;
 import tech.pegasys.ethsigner.core.signing.filebased.FileBasedTransactionSigner;
 
 import java.nio.file.Path;
@@ -29,7 +30,7 @@ import picocli.CommandLine.Spec;
         "This command ensures that received transactions are signed by a key stored in an encrypted file.",
     mixinStandardHelpOptions = true,
     helpCommand = true)
-public class FileBasedTransactionSignerCommand implements Runnable {
+public class FileBasedTransactionSignerCommand extends SignerSubCommand {
 
   @CommandLine.ParentCommand private CommandLineConfig parentCommand;
 
@@ -63,10 +64,7 @@ public class FileBasedTransactionSignerCommand implements Runnable {
   }
 
   @Override
-  public void run() {
-
-    final FileBasedTransactionSigner transactionSigner =
-        new FileBasedTransactionSigner(keyFilePath, passwordFilePath);
-    parentCommand.startEthSigner(transactionSigner);
+  public TransactionSigner createSigner() {
+    return new FileBasedTransactionSigner(keyFilePath, passwordFilePath);
   }
 }
