+++ ethsigner/hashicorpsigner/src/main/java/tech/pegasys/ethsigner/signing/hashicorp/HashicorpTransactionSigner.java
@@ -1,4 +1,4 @@
-/*
+package tech.pegasys.ethsigner.signing.hashicorp;/*
  * Copyright 2019 ConsenSys AG.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
@@ -10,7 +10,6 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.ethsigner.core.signing.hashicorp;
 
 import tech.pegasys.ethsigner.core.signing.CredentialTransactionSigner;
 import tech.pegasys.ethsigner.core.signing.TransactionSignerInitializationException;
