+++ ethsigner/hashicorpsigner/src/test/java/tech/pegasys/ethsigner/signing/hashicorp/HashicorpTransactionSignerTest.java
@@ -10,10 +10,11 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.ethsigner.core.signing.hashicorp;
+package tech.pegasys.ethsigner.signing.hashicorp;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 
+import tech.pegasys.ethsigner.signing.hashicorp.HashicorpTransactionSigner;
 import tech.pegasys.ethsigner.core.signing.TransactionSigner;
 import tech.pegasys.ethsigner.core.signing.TransactionSignerInitializationException;
