+++ ethsigner/subcommandapi/src/main/java/tech/pegasys/ethsigner/ApplicationInfo.java
