+++ ethsigner/subcommandapi/src/main/java/tech/pegasys/ethsigner/CommandLineConfig.java
@@ -13,10 +13,8 @@
 package tech.pegasys.ethsigner;
 
 import tech.pegasys.ethsigner.core.Config;
-import tech.pegasys.ethsigner.core.EthSigner;
 import tech.pegasys.ethsigner.core.signing.ChainIdProvider;
 import tech.pegasys.ethsigner.core.signing.ConfigurationChainId;
-import tech.pegasys.ethsigner.core.signing.TransactionSigner;
 
 import java.io.PrintStream;
 import java.net.InetAddress;
@@ -28,7 +26,6 @@ import com.google.common.base.MoreObjects;
 import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.apache.logging.log4j.core.config.Configurator;
 import picocli.CommandLine;
 import picocli.CommandLine.Command;
 import picocli.CommandLine.Help.Ansi;
@@ -115,55 +112,6 @@ public class CommandLineConfig implements Config {
     this.output = output;
   }
 
-  private HashicorpTransactionSignerCommand hashicorpTransactionSignerConfig;
-
-  private FileBasedTransactionSignerCommand fileBasedTransactionSignerConfig;
-
-  public boolean parse(
-      final CommandLine.IParseResultHandler2<List<Object>> resultHandler, final String... args) {
-
-    commandLine = new CommandLine(this);
-    commandLine.setCaseInsensitiveEnumValuesAllowed(true);
-    commandLine.registerConverter(Level.class, Level::valueOf);
-
-    hashicorpTransactionSignerConfig = new HashicorpTransactionSignerCommand();
-    commandLine.addSubcommand(
-        HashicorpTransactionSignerCommand.COMMAND_NAME, hashicorpTransactionSignerConfig);
-
-    fileBasedTransactionSignerConfig = new FileBasedTransactionSignerCommand();
-    commandLine.addSubcommand(
-        FileBasedTransactionSignerCommand.COMMAND_NAME, fileBasedTransactionSignerConfig);
-
-    // Must manually show the usage/version info, as per the design of picocli
-    // (https://picocli.info/#_printing_help_automatically)
-    try {
-      commandLine.parseWithHandlers(resultHandler, new ExceptionHandler<List<Object>>(), args);
-    } catch (final ParameterException ex) {
-      handleParseException(ex);
-      return false;
-    }
-
-    if (commandLine.isUsageHelpRequested()) {
-      commandLine.usage(output);
-      return false;
-    } else if (commandLine.isVersionHelpRequested()) {
-      commandLine.printVersionHelp(output);
-      return false;
-    }
-    return true;
-  }
-
-  public void handleParseException(final ParameterException ex) {
-    if (logLevel != null && Level.DEBUG.isMoreSpecificThan(logLevel)) {
-      ex.printStackTrace(output);
-    } else {
-      output.println(ex.getMessage());
-    }
-    if (!CommandLine.UnmatchedArgumentException.printSuggestions(ex, output)) {
-      ex.getCommandLine().usage(output, Ansi.AUTO);
-    }
-  }
-
   @Override
   public Level getLogLevel() {
     return logLevel;
@@ -217,38 +165,6 @@ public class CommandLineConfig implements Config {
         .add("chainId", chainId)
         .add("dataDirectory", dataDirectory)
         .add("output", output)
-        .add("hashicorpTransactionSignerConfig", hashicorpTransactionSignerConfig)
-        .add("filebasedTransactionSignerConfig", fileBasedTransactionSignerConfig)
         .toString();
-  }
-
-  public void startEthSigner(final TransactionSigner transactionSigner) {
-    if (transactionSigner == null) {
-      LOG.error("EthSigner cannot be started without a TransactionSigner. Config = {}", this);
-      throw new RuntimeException("EthSigner cannot be started without a TransactionSigner.");
-    }
-    // set log level per CLI flags
-    System.out.println("Setting logging level to " + getLogLevel().name());
-    Configurator.setAllLevels("", getLogLevel());
-
-    LOG.debug("Configuration = {}", this);
-    LOG.info("Version = {}", ApplicationInfo.version());
-
-    final EthSigner signer = new EthSigner(this, transactionSigner);
-    signer.run();
-  }
-
-  private static class ExceptionHandler<R> implements CommandLine.IExceptionHandler2<R> {
-
-    @Override
-    public R handleParseException(final ParameterException ex, final String[] args) {
-      throw new RuntimeException("Exception handled in handleParseException.", ex);
-    }
-
-    @Override
-    public R handleExecutionException(
-        final CommandLine.ExecutionException ex, final CommandLine.ParseResult parseResult) {
-      throw new RuntimeException("Exception handled in handleParseException.", ex);
-    }
   }
 }
