+++ ethsigner/subcommandapi/src/main/java/tech/pegasys/ethsigner/PlatformDetector.java
