+++ ethsigner/subcommandapi/src/main/java/tech/pegasys/ethsigner/VersionInfo.java
