+++ ethsigner/subcommandapi/src/main/java/tech/pegasys/ethsigner/VersionProvider.java
