+++ settings.gradle
@@ -16,3 +16,5 @@ include 'acceptance-tests'
 include 'ethsigner:app'
 include 'ethsigner:core'
 include 'ethsigner:signingapi'
+include 'ethsigner:hashicorpsigner'
+include 'ethsigner:subcommandapi'
